requirements.txt
amqp==5.0.9
asgiref==3.4.1
autopep8==1.6.0
backcall==0.2.0
billiard==3.6.4.0
boto3==1.17.110
botocore==1.20.110
celery==5.2.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.5
django-celery-beat==2.2.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-ses==2.3.1
django-storages==1.12.3
django-timezone-field==4.2.3
djangorestframework==3.12.4
drf-yasg==1.20.0
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
ipdb==0.13.9
ipython==7.30.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.1
jmespath==0.10.0
kombu==5.2.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
numpy==1.22.3
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.0
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.24
psycopg2==2.9.2
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.8.0
Pygments==2.11.1
PyJWT==2.1.0
pyparsing==2.4.7
python-crontab==2.6.0
python-dateutil==2.8.1
python-dotenv==0.20.0
pytz==2021.3
pyxlsb==1.0.9
razorpay==1.2.0
reportlab==3.6.3
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
s3transfer==0.4.2
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.1.1
twilio==7.3.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
